Worker: extract types + telemetry + commentary from index.ts
First Batch-3 route slice. index.ts was ~8,060 lines with everything inline;
the keystone problem is that Bindings lived there, so any extracted module that
needs the env type would cycle back through this entry file.

- src/worker/types.ts: Bindings + JobMessage (neutral module; crons import
  JobMessage from here instead of ./index)
- src/worker/telemetry.ts: TelemetryRecord + classifyError / recordTelemetry /
  logTelemetry / runTelemetryRec (captureLlmUsage + recordObserved* stay in
  index.ts since they depend on RunCtx)
- src/worker/commentary.ts: fetchCommentaryWorks + the two commentary routes via
  registerCommentaryRoutes(app)

Pure move, behavior-preserving (paths, bodies, cache keys, model lists, telemetry
unchanged). index.ts 8,063 -> 7,687. Unblocks further per-domain route slices.
